[refactor] Company API가 Interface를 의존하도록 Refactoring
