+++ module-api/src/main/java/com/welcommu/moduleapi/company/CompanyController.java
@@ -3,10 +3,11 @@ package com.welcommu.moduleapi.company;
 import com.welcommu.modulecommon.dto.ApiResponse;
 import com.welcommu.moduledomain.auth.AuthUserDetailsImpl;
 import com.welcommu.moduledomain.company.Company;
+import com.welcommu.moduleservice.company.CompanyService;
 import com.welcommu.moduleservice.company.dto.CompanyModifyRequest;
 import com.welcommu.moduleservice.company.dto.CompanyRequest;
 import com.welcommu.moduleservice.company.dto.CompanyResponse;
-import com.welcommu.moduleservice.company.CompanyService;
+import com.welcommu.moduleservice.company.CompanyServiceImpl;
 import com.welcommu.moduleservice.user.dto.UserResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
